Refactor detections table to use double precision for area and depth scores

+++ drizzle/schema.sql
@@ -42,8 +42,8 @@ CREATE TABLE "detections" (
     "image_id" integer NOT NULL,
     "bbox" jsonb NOT NULL, -- Store as JSON array [x_min, y_min, x_max, y_max]
     "confidence" double precision NOT NULL,
-        "area_score" integer,
-    "depth_score" integer,
+    "area_score" double precision,
+    "depth_score" double precision,
     "false_positive" integer DEFAULT 0,
     CONSTRAINT "detections_image_id_images_id_fk" FOREIGN KEY ("image_id") REFERENCES "images"("id") ON DELETE CASCADE
 );
